Decoder: Introduce Pre-Loading and Off-Time Queueing
There is now a 32 rectangle queue that is used to try and prepare for
both the first frame of rendering and potential jumping to a frame with
a lot of rectangles in it. It makes the first frame of slow video files
faster, but the extra time gained is diminished quickly with high
framerates, making it indiscernable usually. :s